refactor: split js into albums/main/login.js
