+++ app/ui/index.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>iCloud Photo Sorter</title>
     <link rel="stylesheet" href="css/style.css" />
-    <script defer src="js/login.js"></script>
+    <script type="module" src="js/main.js"></script>
   </head>
   <body>
     <div class="login-container">
